Se termina integracion de diferentes master page en una vista

+++ SoftmatDesk/Controllers/ticketsController.cs
@@ -30,6 +30,20 @@ namespace SoftmatDesk.Controllers
                 ViewBag.Id = Session["id"];
                 return View(tickets);
             }
+            else if (Session["Rol"].ToString() == "Cliente" || Session["Rol"].ToString() == "Client")
+            {
+                int IdC = Int32.Parse(Session["idC"].ToString());
+                var tickets = db.tickets.Where(t => t.Cliente_idCliente == IdC).ToList();
+                ViewBag.Categorias_idCategorias = new SelectList(db.categorias, "idCategorias", "Categoria");
+                ViewBag.Cliente_idCliente = new SelectList(db.cliente, "idCliente", "Razon_social");
+                ViewBag.Nivel_prioridad_idNivel_prioridad = new SelectList(db.nivel_prioridad, "idNivel_prioridad", "Prioridad");
+                ViewBag.Sedes_idSedes = new SelectList(db.sedes, "idSedes", "Nom_Sede");
+                ViewBag.SmUsuarios_idsmUsuarios = new SelectList(db.smusuarios, "idsmUsuarios", "Nombres");
+                ViewBag.Usuario_idUsuario = new SelectList(db.usuario, "idUsuario", "Nombres");
+                ViewBag.Nombre = Session["Sesion"];
+                ViewBag.Id = Session["id"];
+                return View(tickets);
+            }
 
             return Content("No tiene acceso");
         }
@@ -121,16 +135,90 @@ namespace SoftmatDesk.Controllers
         public ActionResult Create()
         {
 
-            ViewBag.Categorias_idCategorias = new SelectList(db.categorias, "idCategorias", "Categoria");
-            ViewBag.Cliente_idCliente = new SelectList(db.cliente, "idCliente", "Razon_social");
-            ViewBag.Nivel_prioridad_idNivel_prioridad = new SelectList(db.nivel_prioridad, "idNivel_prioridad", "Prioridad");
-            ViewBag.Sedes_idSedes = new SelectList(db.sedes, "idSedes", "Nom_Sede");
-            ViewBag.SmUsuarios_idsmUsuarios = new SelectList(db.smusuarios, "idsmUsuarios", "Nombres");
-            ViewBag.Usuario_idUsuario = new SelectList(db.usuario, "idUsuario", "Nombres");
-            return View();
+            if (Session["Sesion"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            else if(Session["Rol"].ToString() == "Administrador" || Session["Rol"].ToString() == "Admin") {
+                ViewBag.Categorias_idCategorias = new SelectList(db.categorias, "idCategorias", "Categoria");
+                ViewBag.Cliente_idCliente = new SelectList(db.cliente, "idCliente", "Razon_social");
+                ViewBag.Nivel_prioridad_idNivel_prioridad = new SelectList(db.nivel_prioridad, "idNivel_prioridad", "Prioridad");
+                ViewBag.Sedes_idSedes = new SelectList(db.sedes, "idSedes", "Nom_Sede");
+                ViewBag.SmUsuarios_idsmUsuarios = new SelectList(db.smusuarios, "idsmUsuarios", "Nombres");
+                ViewBag.Usuario_idUsuario = new SelectList(db.usuario, "idUsuario", "Nombres");
+                return View();
+            }
+            return View("Acceso no autorizado");
 
         }
 
+        public ActionResult CreateUs()
+        {
+            if (Session["Sesion"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            else if(Session["Rol"].ToString() == "Usuario" || Session["Rol"].ToString() == "User"){
+                var Cliente = db.cliente.Where(c => c.idCliente == Int32.Parse(Session["idC"].ToString())).ToList();
+                ViewBag.Categorias_idCategorias = new SelectList(db.categorias, "idCategorias", "Categoria");
+                //ViewBag.Cliente_idCliente = new SelectList(db.cliente, "idCliente", "Razon_social");
+                ViewBag.Cliente_idCliente = new SelectList(Cliente, "idCliente", "Razon_social");
+                ViewBag.Nivel_prioridad_idNivel_prioridad = new SelectList(db.nivel_prioridad, "idNivel_prioridad", "Prioridad");
+                var Sedes = db.sedes.Where(s => s.Cliente_idCliente == Int32.Parse(Session["idC"].ToString())).ToList();
+                //ViewBag.Sedes_idSedes = new SelectList(db.sedes, "idSedes", "Nom_Sede");
+                ViewBag.Sedes_idSedes = new SelectList(Sedes, "idSedes", "Nom_Sede");
+                ViewBag.SmUsuarios_idsmUsuarios = new SelectList(db.smusuarios, "idsmUsuarios", "Nombres");
+                var Usuario = db.usuario.Where(u => u.idUsuario == Int32.Parse(Session["id"].ToString())).ToList();
+                //ViewBag.Usuario_idUsuario = new SelectList(db.usuario, "idUsuario", "Nombres");
+                ViewBag.Usuario_idUsuario = new SelectList(Usuario, "idUsuario", "Nombres");
+                return View();
+            }
+            return View("Acceso no autorizado");
+
+        }
+
+        public ActionResult CreateSop()
+        {
+            if (Session["Sesion"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            else if(Session["Rol"].ToString() == "Soporte" || Session["Rol"].ToString() == "Sop") {
+                ViewBag.Categorias_idCategorias = new SelectList(db.categorias, "idCategorias", "Categoria");
+                ViewBag.Cliente_idCliente = new SelectList(db.cliente, "idCliente", "Razon_social");
+                ViewBag.Nivel_prioridad_idNivel_prioridad = new SelectList(db.nivel_prioridad, "idNivel_prioridad", "Prioridad");
+                ViewBag.Sedes_idSedes = new SelectList(db.sedes, "idSedes", "Nom_Sede");
+                ViewBag.SmUsuarios_idsmUsuarios = new SelectList(db.smusuarios, "idsmUsuarios", "Nombres");
+                ViewBag.Usuario_idUsuario = new SelectList(db.usuario, "idUsuario", "Nombres");
+                return View();
+            }
+            return View("Acceso no autorizado");
+        }
+
+        public ActionResult CreateCl()
+        {
+            if (Session["Sesion"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            else if(Session["Rol"].ToString() == "Cliente" || Session["Rol"].ToString() == "Client") {
+                    ViewBag.Categorias_idCategorias = new SelectList(db.categorias, "idCategorias", "Categoria");
+                    var Cliente = db.cliente.Where(c => c.idCliente == Int32.Parse(Session["idC"].ToString())).ToList();
+                    //ViewBag.Cliente_idCliente = new SelectList(db.cliente, "idCliente", "Razon_social");
+                    ViewBag.Cliente_idCliente = new SelectList(Cliente, "idCliente", "Razon_social");
+                    ViewBag.Nivel_prioridad_idNivel_prioridad = new SelectList(db.nivel_prioridad, "idNivel_prioridad", "Prioridad");
+                    var Sedes = db.sedes.Where(s => s.Cliente_idCliente == Int32.Parse(Session["idC"].ToString())).ToList();
+                    //ViewBag.Sedes_idSedes = new SelectList(db.sedes, "idSedes", "Nom_Sede");
+                    ViewBag.Sedes_idSedes = new SelectList(Sedes, "idSedes", "Nom_Sede");
+                    ViewBag.SmUsuarios_idsmUsuarios = new SelectList(db.smusuarios, "idsmUsuarios", "Nombres");
+                    var Usuario = db.usuario.Where(u => u.Cliente_idCliente == Int32.Parse(Session["idC"].ToString())).ToList();
+                    //ViewBag.Usuario_idUsuario = new SelectList(db.usuario, "idUsuario", "Nombres");
+                    ViewBag.Usuario_idUsuario = new SelectList(Usuario, "idUsuario", "Nombres");
+                    return View();
+                }
+            return View("Acceso no autorizado");
+
+        }
         // POST: tickets/Create
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
@@ -138,7 +226,11 @@ namespace SoftmatDesk.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Create([Bind(Include = "NoTickets,Categorias_idCategorias,Usuario_idUsuario,Nivel_prioridad_idNivel_prioridad,Cliente_idCliente,Sedes_idSedes,Descripcion_falla,Apertura,Cierre,SmUsuarios_idsmUsuarios")] tickets tickets)
         {
-            if (ModelState.IsValid)
+            if (Session["Sesion"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            else if(ModelState.IsValid)
             {
                 db.tickets.Add(tickets);
                 await db.SaveChangesAsync();
@@ -151,11 +243,87 @@ namespace SoftmatDesk.Controllers
             ViewBag.Sedes_idSedes = new SelectList(db.sedes, "idSedes", "Nom_Sede", tickets.Sedes_idSedes);
             ViewBag.SmUsuarios_idsmUsuarios = new SelectList(db.smusuarios, "idsmUsuarios", "Nombres", tickets.SmUsuarios_idsmUsuarios);
             ViewBag.Usuario_idUsuario = new SelectList(db.usuario, "idUsuario", "Nombres", tickets.Usuario_idUsuario);
+            if(Session["Rol"].ToString() =="Administrador" || Session["Session"].ToString() == "Admin")
+            {
+                return View("tikets");
+            }
+            else if (Session["Rol"].ToString() == "Soporte" || Session["Session"].ToString() == "Sop")
+            {
+                return View("ListSop");
+            }
+            else if (Session["Rol"].ToString() == "Cliente" || Session["Session"].ToString() == "Client")
+            {
+                return View("ListCl");
+            }
+            else if (Session["Rol"].ToString() == "Usuario" || Session["Session"].ToString() == "User")
+            {
+                return View("ListUs");
+            }
             return View(tickets);
         }
 
         // GET: tickets/Edit
         public async Task<ActionResult> Edit(int? id)
+        {
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            tickets tickets = await db.tickets.FindAsync(id);
+            if (tickets == null)
+            {
+                return HttpNotFound();
+            }
+            ViewBag.Categorias_idCategorias = new SelectList(db.categorias, "idCategorias", "Categoria", tickets.Categorias_idCategorias);
+            ViewBag.Cliente_idCliente = new SelectList(db.cliente, "idCliente", "Razon_social", tickets.Cliente_idCliente);
+            ViewBag.Nivel_prioridad_idNivel_prioridad = new SelectList(db.nivel_prioridad, "idNivel_prioridad", "Prioridad", tickets.Nivel_prioridad_idNivel_prioridad);
+            ViewBag.Sedes_idSedes = new SelectList(db.sedes, "idSedes", "Nom_Sede", tickets.Sedes_idSedes);
+            ViewBag.SmUsuarios_idsmUsuarios = new SelectList(db.smusuarios, "idsmUsuarios", "Nombres", tickets.SmUsuarios_idsmUsuarios);
+            ViewBag.Usuario_idUsuario = new SelectList(db.usuario, "idUsuario", "Nombres", tickets.Usuario_idUsuario);
+            return View(tickets);
+        }
+
+        public async Task<ActionResult> EditSop(int? id)
+        {
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            tickets tickets = await db.tickets.FindAsync(id);
+            if (tickets == null)
+            {
+                return HttpNotFound();
+            }
+            ViewBag.Categorias_idCategorias = new SelectList(db.categorias, "idCategorias", "Categoria", tickets.Categorias_idCategorias);
+            ViewBag.Cliente_idCliente = new SelectList(db.cliente, "idCliente", "Razon_social", tickets.Cliente_idCliente);
+            ViewBag.Nivel_prioridad_idNivel_prioridad = new SelectList(db.nivel_prioridad, "idNivel_prioridad", "Prioridad", tickets.Nivel_prioridad_idNivel_prioridad);
+            ViewBag.Sedes_idSedes = new SelectList(db.sedes, "idSedes", "Nom_Sede", tickets.Sedes_idSedes);
+            ViewBag.SmUsuarios_idsmUsuarios = new SelectList(db.smusuarios, "idsmUsuarios", "Nombres", tickets.SmUsuarios_idsmUsuarios);
+            ViewBag.Usuario_idUsuario = new SelectList(db.usuario, "idUsuario", "Nombres", tickets.Usuario_idUsuario);
+            return View(tickets);
+        }
+
+        public async Task<ActionResult> EditCl(int? id)
+        {
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            tickets tickets = await db.tickets.FindAsync(id);
+            if (tickets == null)
+            {
+                return HttpNotFound();
+            }
+            ViewBag.Categorias_idCategorias = new SelectList(db.categorias, "idCategorias", "Categoria", tickets.Categorias_idCategorias);
+            ViewBag.Cliente_idCliente = new SelectList(db.cliente, "idCliente", "Razon_social", tickets.Cliente_idCliente);
+            ViewBag.Nivel_prioridad_idNivel_prioridad = new SelectList(db.nivel_prioridad, "idNivel_prioridad", "Prioridad", tickets.Nivel_prioridad_idNivel_prioridad);
+            ViewBag.Sedes_idSedes = new SelectList(db.sedes, "idSedes", "Nom_Sede", tickets.Sedes_idSedes);
+            ViewBag.SmUsuarios_idsmUsuarios = new SelectList(db.smusuarios, "idsmUsuarios", "Nombres", tickets.SmUsuarios_idsmUsuarios);
+            ViewBag.Usuario_idUsuario = new SelectList(db.usuario, "idUsuario", "Nombres", tickets.Usuario_idUsuario);
+            return View(tickets);
+        }
+
+        public async Task<ActionResult> EditUs(int? id)
         {
             if (id == null)
             {
@@ -182,7 +350,11 @@ namespace SoftmatDesk.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Edit([Bind(Include = "NoTickets,Categorias_idCategorias,Usuario_idUsuario,Nivel_prioridad_idNivel_prioridad,Cliente_idCliente,Sedes_idSedes,Descripcion_falla,Apertura,Cierre,SmUsuarios_idsmUsuarios")] tickets tickets)
         {
-            if (ModelState.IsValid)
+            if (Session["Sesion"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            else if(ModelState.IsValid)
             {
                 db.Entry(tickets).State = EntityState.Modified;
                 await db.SaveChangesAsync();
@@ -194,6 +366,22 @@ namespace SoftmatDesk.Controllers
             ViewBag.Sedes_idSedes = new SelectList(db.sedes, "idSedes", "Nom_Sede", tickets.Sedes_idSedes);
             ViewBag.SmUsuarios_idsmUsuarios = new SelectList(db.smusuarios, "idsmUsuarios", "Nombres", tickets.SmUsuarios_idsmUsuarios);
             ViewBag.Usuario_idUsuario = new SelectList(db.usuario, "idUsuario", "Nombres", tickets.Usuario_idUsuario);
+            if (Session["Rol"].ToString() == "Administrador" || Session["Session"].ToString() == "Admin")
+            {
+                return View("tikets");
+            }
+            else if (Session["Rol"].ToString() == "Soporte" || Session["Session"].ToString() == "Sop")
+            {
+                return View("ListSop");
+            }
+            else if (Session["Rol"].ToString() == "Cliente" || Session["Session"].ToString() == "Client")
+            {
+                return View("ListCl");
+            }
+            else if (Session["Rol"].ToString() == "Usuario" || Session["Session"].ToString() == "User")
+            {
+                return View("ListUs");
+            }
             return View(tickets);
         }
 
@@ -217,9 +405,29 @@ namespace SoftmatDesk.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> DeleteConfirmed(int id)
         {
-            tickets tickets = await db.tickets.FindAsync(id);
+            if (Session["Sesion"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+                tickets tickets = await db.tickets.FindAsync(id);
             db.tickets.Remove(tickets);
             await db.SaveChangesAsync();
+            if (Session["Rol"].ToString() == "Administrador" || Session["Session"].ToString() == "Admin")
+            {
+                return View("tikets");
+            }
+            else if (Session["Rol"].ToString() == "Soporte" || Session["Session"].ToString() == "Sop")
+            {
+                return View("ListSop");
+            }
+            else if (Session["Rol"].ToString() == "Cliente" || Session["Session"].ToString() == "Client")
+            {
+                return View("ListCl");
+            }
+            else if (Session["Rol"].ToString() == "Usuario" || Session["Session"].ToString() == "User")
+            {
+                return View("ListUs");
+            }
             return RedirectToAction("Index");
         }
